Move & restructrue eventmaster http.Handler
In preparation for adding tests for the http routes (to validate new
grafana code) this change changes eventmaster.Server to implement
http.Handler.

It consolidates route registration code into a single, orthogonal
(swappable) function. It also reduces the layers of abstraction by one,
by removing the NewHTTPServer function.
